+++ src/imio/pm/wsclient/browser/__init__.py
@@ -1,1 +1,1 @@
-#I'm a python package
+# I'm a python package