+++ src/imio/pm/wsclient/testing.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
+
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
-from plone.app.testing import PloneWithPackageLayer
 from plone.testing import z2
 from plone.testing import zca
 from Products.PloneMeeting.testing import PMLayer