+++ src/imio/pm/wsclient/tests/testEvents.py
@@ -2,24 +2,7 @@
 #
 # File: testEvents.py
 #
-# Copyright (c) 2012 by CommunesPlone
-#
 # GNU General Public License (GPL)
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
-# 02110-1301, USA.
 #
 
 from imio.pm.wsclient.interfaces import ISentToPMEvent
@@ -90,7 +73,7 @@ class testEvents(WS4PMCLIENTTestCase):
         gsm = getGlobalSiteManager()
         gsm.registerHandler(sent_to_pm_handler, (Interface, ISentToPMEvent))
 
-        #send the element to pm, the event handler should raise the exception
+        # send the element to pm, the event handler should raise the exception
         with self.assertRaises(SentToPm):
             self.sending_view._doSendToPloneMeeting()