+++ src/imio/pm/wsclient/tests/testSettings.py
@@ -2,24 +2,7 @@
 #
 # File: testItemMethods.py
 #
-# Copyright (c) 2012 by CommunesPlone
-#
 # GNU General Public License (GPL)
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
-# 02110-1301, USA.
 #
 
 from AccessControl import Unauthorized
@@ -36,10 +19,10 @@ from zope.event import notify
 from zope.tales.tales import CompilerError
 from zope.traversing.interfaces import BeforeTraverseEvent
 
-import unittest2
+import unittest
 
 
-class testSettings(unittest2.TestCase):
+class testSettings(unittest.TestCase):
     """
         Tests the browser.settings SOAP client methods
     """